pyproject.toml
[tool.poetry]
name = "charsheet-site"
version = "0.1.0"
description = ""
authors = ["DimitryRukhadze <lokiinflames@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
Django = "4.2.6"
django-mathfilters = "1.0.0"
djangorestframework = "^3.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
